refactor: extract widget collection service
