+++ app/src/main/java/com/kgs/calendar/widget/KgsWidgetProvider.kt
@@ -135,7 +135,7 @@ internal const val WIDGET_TASK_SUBTASK_ARROW_STROKE_DP = WidgetTaskCardRenderer.
 internal val WIDGET_TASK_CARD_RENDERER = WidgetTaskCardRenderer()
 private const val WIDGET_TASKS_PRIORITY_FRAME_COUNT = 20
 private const val WIDGET_AGENDA_PRIORITY_FRAME_COUNT = 30
-private const val WIDGET_COLLECTION_VIEW_TYPE_COUNT = 8
+internal const val WIDGET_COLLECTION_VIEW_TYPE_COUNT = 8
 internal const val WIDGET_AGENDA_ART_BITMAP_SCALE = 1.15f
 internal const val WIDGET_TASK_PRIORITY_BITMAP_SCALE = 1.15f
 internal const val WIDGET_TASK_TRANSITION_BITMAP_SCALE = 0.78f
@@ -4906,151 +4906,4 @@ internal class KgsWidgetRenderer(
             PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE,
         )
     }
-}
-
-class KgsWidgetCollectionService : RemoteViewsService() {
-    override fun onGetViewFactory(intent: Intent): RemoteViewsFactory {
-        val kind = intent.getStringExtra(EXTRA_WIDGET_KIND)
-            ?.let { runCatching { KgsWidgetKind.valueOf(it) }.getOrNull() }
-            ?: KgsWidgetKind.Agenda
-        val appWidgetId = intent.getIntExtra(AppWidgetManager.EXTRA_APPWIDGET_ID, AppWidgetManager.INVALID_APPWIDGET_ID)
-        return if (kind == KgsWidgetKind.Day) {
-            KgsWidgetDayCollectionFactory(applicationContext, appWidgetId)
-        } else {
-            KgsWidgetCollectionFactory(applicationContext, kind, appWidgetId)
-        }
-    }
-}
-
-private class KgsWidgetDayCollectionFactory(
-    private val context: Context,
-    private val appWidgetId: Int,
-) : RemoteViewsService.RemoteViewsFactory {
-    private val packageName = context.packageName
-    private var rows: List<WidgetDayGridRow> = emptyList()
-    private var rowViews: List<RemoteViews> = emptyList()
-
-    override fun onCreate() = Unit
-
-    override fun onDataSetChanged() {
-        val startedAt = SystemClock.elapsedRealtime()
-        runCatching {
-            val snapshot = runBlocking {
-                KgsWidgetRenderer(context).dayGridCollectionSnapshot(appWidgetId)
-            }
-            val textContext = context.withWidgetLocale(snapshot.settings.locale)
-            rows = snapshot.rows
-            rowViews = rows.map { row ->
-                row.toRemoteViews(
-                    context = textContext,
-                    packageName = packageName,
-                    palette = snapshot.palette,
-                    widthDp = snapshot.widthDp,
-                    appWidgetId = appWidgetId,
-                )
-            }
-        }.onFailure { error ->
-            Log.e(TAG, "Failed to load Day widget grid rows", error)
-            rows = emptyList()
-            rowViews = emptyList()
-        }
-        WidgetLog.d(context, "Loaded ${rows.size} Day widget grid rows in ${SystemClock.elapsedRealtime() - startedAt}ms")
-    }
-
-    override fun onDestroy() {
-        rows = emptyList()
-        rowViews = emptyList()
-    }
-
-    override fun getCount(): Int = rowViews.size
-
-    override fun getViewAt(position: Int): RemoteViews {
-        return rowViews.getOrNull(position)
-            ?: RemoteViews(packageName, R.layout.widget_collection_spacer)
-    }
-
-    override fun getLoadingView(): RemoteViews = RemoteViews(packageName, R.layout.widget_collection_spacer)
-
-    override fun getViewTypeCount(): Int = 2
-
-    override fun getItemId(position: Int): Long = rows.getOrNull(position)?.stableId ?: position.toLong()
-
-    override fun hasStableIds(): Boolean = true
-}
-
-private class KgsWidgetCollectionFactory(
-    private val context: Context,
-    private val kind: KgsWidgetKind,
-    private val appWidgetId: Int,
-    private val zoneId: ZoneId = ZoneId.systemDefault(),
-) : RemoteViewsService.RemoteViewsFactory {
-    private val packageName = context.packageName
-    private var rows: List<WidgetListRow> = emptyList()
-    private var rowViews: List<RemoteViews> = emptyList()
-    private var settings = WidgetRenderSettings()
-    private var palette = WidgetPalette.from(context, AppThemeMode.KgsBlue, AppColorMode.Auto)
-    private var renderOptions = WidgetCollectionRenderOptions(
-        taskArtWidthDp = WidgetSize.from(
-            context,
-            AppWidgetManager.getInstance(context).getAppWidgetOptions(appWidgetId),
-            kind,
-        ).collectionArtWidthDp(kind),
-    )
-
-    override fun onCreate() = Unit
-
-    override fun onDataSetChanged() {
-        val startedAt = SystemClock.elapsedRealtime()
-        runCatching {
-            val cached = KgsWidgetCollectionRowsCache.get(kind, appWidgetId)
-            if (cached != null) {
-                settings = cached.settings
-                palette = cached.palette
-                rows = cached.rows
-                renderOptions = renderOptions.withTaskArtWidth(cached.taskArtWidthDp)
-            } else {
-                runBlocking {
-                    val dataSource = KgsWidgetDataSource(context, zoneId)
-                    settings = dataSource.loadSettings(kind)
-                    palette = WidgetPalette.from(context, settings.themeMode, settings.colorMode)
-                    rows = dataSource.listRows(kind, settings, appWidgetId)
-                    renderOptions = renderOptions.withTaskArtWidth(
-                        WidgetSize.from(
-                            context,
-                            AppWidgetManager.getInstance(context).getAppWidgetOptions(appWidgetId),
-                            kind,
-                        ).collectionArtWidthDp(kind),
-                    )
-                }
-            }
-            val textContext = context.withWidgetLocale(settings.locale)
-            rowViews = rows.map { row ->
-                row.toRemoteViews(textContext, packageName, palette, kind, appWidgetId, renderOptions)
-            }
-        }.onFailure { error ->
-            Log.e(TAG, "Failed to load ${kind.name} widget rows", error)
-            rows = emptyList()
-            rowViews = emptyList()
-        }
-        WidgetLog.d(context, "Loaded ${rows.size} ${kind.name} widget rows in ${SystemClock.elapsedRealtime() - startedAt}ms")
-    }
-
-    override fun onDestroy() {
-        rows = emptyList()
-        rowViews = emptyList()
-    }
-
-    override fun getCount(): Int = rows.size
-
-    override fun getViewAt(position: Int): RemoteViews =
-        rowViews.getOrNull(position)
-            ?: RemoteViews(packageName, R.layout.widget_collection_spacer)
-
-    override fun getLoadingView(): RemoteViews = RemoteViews(packageName, R.layout.widget_collection_spacer)
-
-    override fun getViewTypeCount(): Int = WIDGET_COLLECTION_VIEW_TYPE_COUNT
-
-    override fun getItemId(position: Int): Long = rows.getOrNull(position)?.stableId ?: position.toLong()
-
-    override fun hasStableIds(): Boolean = true
 }
